+++ lib/components/tree-node-children.component.ts
@@ -1,9 +1,8 @@
-import { Component, Input, ViewEncapsulation, ElementRef } from '@angular/core';
+import { Component, Input, ViewEncapsulation } from '@angular/core';
 import { TreeNode } from '../models/tree-node.model';
-import { deprecatedSelector } from '../deprecated-selector';
 
 @Component({
-  selector: 'TreeNodeChildren, tree-node-children',
+  selector: 'tree-node-children',
   encapsulation: ViewEncapsulation.None,
   styles: [
     '.tree-children.tree-children-no-padding { padding-left: 0 }',
@@ -38,8 +37,4 @@ import { deprecatedSelector } from '../deprecated-selector';
 export class TreeNodeChildrenComponent {
   @Input() node: TreeNode;
   @Input() templates: any;
-
-  constructor(private elementRef: ElementRef) {
-    deprecatedSelector('TreeNodeChildren', 'tree-node-children', elementRef);
-  }
 }