+++ lib/components/tree-node-content.component.ts
@@ -1,9 +1,8 @@
-import { Component, Input, ViewEncapsulation, TemplateRef, ElementRef } from '@angular/core';
+import { Component, Input, ViewEncapsulation, TemplateRef } from '@angular/core';
 import { TreeNode } from '../models/tree-node.model';
-import { deprecatedSelector } from '../deprecated-selector';
 
 @Component({
-  selector: 'TreeNodeContent, tree-node-content',
+  selector: 'tree-node-content',
   encapsulation: ViewEncapsulation.None,
   template: `
   <span *ngIf="!template">{{ node.displayField }}</span>
@@ -16,8 +15,4 @@ export class TreeNodeContent {
   @Input() node: TreeNode;
   @Input() index: number;
   @Input() template: TemplateRef<any>;
-
-  constructor(private elementRef: ElementRef) {
-    deprecatedSelector('TreeNodeContent', 'tree-node-content', elementRef);
-  }
 }