+++ lib/components/tree-node-drop-slot.component.ts
@@ -1,6 +1,5 @@
-import { Component, Input, ViewEncapsulation, ElementRef } from '@angular/core';
+import { Component, Input, ViewEncapsulation } from '@angular/core';
 import { TreeNode } from '../models/tree-node.model';
-import { deprecatedSelector } from '../deprecated-selector';
 
 @Component({
   selector: 'TreeNodeDropSlot, tree-node-drop-slot',
@@ -20,10 +19,6 @@ import { deprecatedSelector } from '../deprecated-selector';
 export class TreeNodeDropSlot {
   @Input() node: TreeNode;
   @Input() dropIndex: number;
-
-  constructor(private elementRef: ElementRef) {
-    deprecatedSelector('TreeNodeDropSlot', 'tree-node-drop-slot', elementRef);
-  }
 
   onDrop($event) {
     this.node.mouseAction('drop', $event.event, {